backend/requirements.txt
# Core Django
Django==5.2.6
djangorestframework==3.16.1

# Database
psycopg2-binary==2.9.10

# Environment and Configuration
python-decouple==3.8

# API Features
django-cors-headers==4.9.0
django-filter==25.1
drf-yasg==1.21.11

# Required by drf-yasg
inflection==0.5.1
packaging==25.0
pytz==2025.2
PyYAML==6.0.2
uritemplate==4.2.0

# Core dependencies
asgiref==3.9.2
sqlparse==0.5.3
